+++ cli/src/cli_util.rs
@@ -161,7 +161,6 @@ use crate::command_error::config_error_with_message;
 use crate::command_error::handle_command_result;
 use crate::command_error::internal_error;
 use crate::command_error::internal_error_with_message;
-use crate::command_error::print_error_sources;
 use crate::command_error::print_parse_diagnostics;
 use crate::command_error::user_error;
 use crate::command_error::user_error_with_message;
@@ -1237,6 +1236,8 @@ pub struct WorkspaceCommandHelper {
     // TODO: Parsed template can be cached if it doesn't capture 'repo lifetime
     commit_summary_template_text: String,
     op_summary_template_text: String,
+    #[cfg(feature = "git")]
+    fetch_summary_template_text: String,
     may_snapshot_working_copy: bool,
     may_update_working_copy: bool,
 }
@@ -1274,6 +1275,8 @@ impl WorkspaceCommandHelper {
         let settings = workspace.settings();
         let commit_summary_template_text = settings.get_string("templates.commit_summary")?;
         let op_summary_template_text = settings.get_string("templates.op_summary")?;
+        #[cfg(feature = "git")]
+        let fetch_summary_template_text = settings.get_string("templates.git_fetch")?;
         let may_update_working_copy =
             may_snapshot_working_copy && env.command.should_commit_transaction();
 
@@ -1282,6 +1285,8 @@ impl WorkspaceCommandHelper {
             user_repo: ReadonlyUserRepo::new(repo),
             env,
             commit_summary_template_text,
+            #[cfg(feature = "git")]
+            fetch_summary_template_text,
             op_summary_template_text,
             may_snapshot_working_copy,
             may_update_working_copy,
@@ -1290,6 +1295,8 @@ impl WorkspaceCommandHelper {
         // mutable operation.
         helper.parse_operation_template(ui, &helper.op_summary_template_text)?;
         helper.parse_commit_template(ui, &helper.commit_summary_template_text)?;
+        #[cfg(feature = "git")]
+        helper.parse_refstatus_template(ui, &helper.fetch_summary_template_text)?;
         helper.parse_commit_template(ui, SHORT_CHANGE_ID_TEMPLATE_TEXT)?;
         Ok(helper)
     }
@@ -1509,6 +1516,11 @@ impl WorkspaceCommandHelper {
 
     pub fn env(&self) -> &WorkspaceCommandEnvironment {
         &self.env
+    }
+
+    #[cfg(feature = "git")]
+    pub fn fetch_summary_template_text(&self) -> &str {
+        &self.fetch_summary_template_text
     }
 
     async fn prepare_working_copy_mutation(&self) -> Result<Commit, CommandError> {
@@ -1931,7 +1943,7 @@ to the current parents may contain changes from multiple commits.
     }
 
     /// Parses template that is validated by `Self::new()`.
-    fn reparse_valid_template<'a, C, L>(
+    pub(crate) fn reparse_valid_template<'a, C, L>(
         &self,
         language: &L,
         template_text: &str,
@@ -1956,6 +1968,17 @@ to the current parents may contain changes from multiple commits.
         ui: &Ui,
         template_text: &str,
     ) -> Result<TemplateRenderer<'_, Commit>, CommandError> {
+        let language = self.commit_template_language();
+        self.parse_template(ui, &language, template_text)
+    }
+
+    /// Parses refstatus template into evaluation tree.
+    #[cfg(feature = "git")]
+    pub fn parse_refstatus_template(
+        &self,
+        ui: &Ui,
+        template_text: &str,
+    ) -> Result<TemplateRenderer<'_, crate::git_util::RefStatus>, CommandError> {
         let language = self.commit_template_language();
         self.parse_template(ui, &language, template_text)
     }
@@ -2741,7 +2764,7 @@ async fn try_reset_git_head(
         Ok(()) => Ok(()),
         Err(err @ jj_lib::git::GitResetHeadError::UpdateHeadRef(_)) => {
             writeln!(ui.warning_default(), "{err}")?;
-            print_error_sources(ui, err.source())?;
+            crate::command_error::print_error_sources(ui, err.source())?;
             Ok(())
         }
         Err(err) => Err(err.into()),